rivers/kernel_test: rename testcase to kernel_test
This driver is to support testing kernel functionalities from user apps.
The start of this module was supporting kernel testcase so that it was named
testcase. But currently this is used to test other format of kernel test.
Let's rename testcase to kernel_test.

This commit include some minor tidy-up like reordering the headers.
