+++ apps/examples/mpu/mpu_test.c
@@ -24,7 +24,7 @@
 #include <semaphore.h>
 #include <fcntl.h>
 #include <sys/ioctl.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 #include <tinyara/mpu_test.h>
 
 #ifdef CONFIG_BUILD_KERNEL
@@ -38,9 +38,9 @@ int mpu_tc_main(int argc, char *argv[])
 	struct mputest_arg_s obj;
 	int ret;
 
-	tc_fd = open(TESTCASE_DRVPATH, O_WRONLY);
+	tc_fd = open(KERNEL_TEST_DRVPATH, O_WRONLY);
 	if (tc_fd < 0) {
-		printf("\nFAILED to open testcase driver %s\n", TESTCASE_DRVPATH);
+		printf("\nFAILED to open kernel test driver %s\n", KERNEL_TEST_DRVPATH);
 		return ERROR;
 	}
