+++ apps/examples/testcase/le_tc/drivers/Kconfig
@@ -6,7 +6,7 @@
 menuconfig EXAMPLES_TESTCASE_DRIVERS
 	bool "Drivers TestCase Example"
 	default n
-	select TESTCASE_DRV
+	select DRIVERS_KERNEL_TEST
 	---help---
 		Enable the Drivers TestCase Example
