+++ apps/examples/testcase/le_tc/kernel/Kconfig
@@ -6,7 +6,7 @@
 menuconfig EXAMPLES_TESTCASE_KERNEL
 	bool "Kernel TestCase Example"
 	default y
-	select TESTCASE_DRV
+	select DRIVERS_KERNEL_TEST
 	---help---
 		Enable the Kernel TestCase Example
