+++ apps/examples/testcase/le_tc/kernel/kernel_tc_main.c
@@ -24,7 +24,7 @@
 #include <stdio.h>
 #include <errno.h>
 #include <fcntl.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 #include <tinyara/fs/fs.h>
 #include <tinyara/fs/ioctl.h>
 #include "tc_common.h"
@@ -47,9 +47,9 @@ int tc_kernel_main(int argc, char *argv[])
 		return ERROR;
 	}
 
-	g_tc_fd = open(TESTCASE_DRVPATH, O_WRONLY);
+	g_tc_fd = open(KERNEL_TEST_DRVPATH, O_WRONLY);
 	if (g_tc_fd < 0) {
-		tckndbg("Failed to open testcase driver %d\n", errno);
+		tckndbg("Failed to open kernel test driver %d\n", errno);
 		return ERROR;
 	}
