+++ apps/examples/testcase/le_tc/kernel/tc_clock.c
@@ -28,7 +28,7 @@
 #include <tinyara/time.h>
 #include <sys/time.h>
 #include <sys/ioctl.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 #include "tc_internal.h"
 
 const long double l_day = 86400;
