+++ apps/examples/testcase/le_tc/kernel/tc_tash_heapinfo.c
@@ -34,7 +34,7 @@
 #include <sys/ioctl.h>
 #include <apps/shell/tash.h>
 #include <tinyara/mm/mm.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 
 #include "tc_internal.h"
