+++ apps/examples/testcase/le_tc/kernel/tc_task.c
@@ -34,7 +34,7 @@
 #include <unistd.h>
 #include <stdbool.h>
 #include <tinyara/sched.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 #include "tc_internal.h"
 #include <semaphore.h>
