+++ os/drivers/kernel_test/Kconfig
@@ -3,8 +3,8 @@
 # see kconfig-language at https://www.kernel.org/doc/Documentation/kbuild/kconfig-language.txt
 #
 
-config TESTCASE_DRV
+config DRIVERS_KERNEL_TEST
 	bool
 	default n
 	---help---
-		Enables overall support for testcase driver library.
+		Enables overall support for kernel test driver library.
