+++ os/drivers/kernel_test/kernel_test_drv.c
@@ -23,15 +23,16 @@
 #include <tinyara/config.h>
 #include <errno.h>
 #include <debug.h>
+
 #include <tinyara/fs/fs.h>
-#include <tinyara/testcase_drv.h>
+#include <tinyara/kernel_test_drv.h>
 #include <tinyara/sched.h>
 #include "signal/signal.h"
-#include "testcase_proto.h"
+#include "kernel_test_proto.h"
 #ifdef CONFIG_EXAMPLES_MPU_TEST
-#include "binary_manager/binary_manager.h"
 #include <tinyara/binfmt/binfmt.h>
 #include <tinyara/mpu_test.h>
+#include "binary_manager/binary_manager.h"
 #endif
 
 /****************************************************************************
@@ -46,21 +47,21 @@ extern uint32_t _sdata;
  * Private Function Prototypes
  ****************************************************************************/
 
-static int testcase_drv_ioctl(FAR struct file *filep, int cmd, unsigned long arg);
-static ssize_t testcase_drv_read(FAR struct file *filep, FAR char *buffer, size_t len);
-static ssize_t testcase_drv_write(FAR struct file *filep, FAR const char *buffer, size_t len);
+static int kernel_test_drv_ioctl(FAR struct file *filep, int cmd, unsigned long arg);
+static ssize_t kernel_test_drv_read(FAR struct file *filep, FAR char *buffer, size_t len);
+static ssize_t kernel_test_drv_write(FAR struct file *filep, FAR const char *buffer, size_t len);
 
 /****************************************************************************
  * Private Data
  ****************************************************************************/
 
-static const struct file_operations testcase_drv_fops = {
+static const struct file_operations kernel_test_drv_fops = {
 	0,						/* open */
 	0,						/* close */
-	testcase_drv_read,				/* read */
-	testcase_drv_write,				/* write */
+	kernel_test_drv_read,				/* read */
+	kernel_test_drv_write,				/* write */
 	0,						/* seek */
-	testcase_drv_ioctl				/* ioctl */
+	kernel_test_drv_ioctl				/* ioctl */
 #ifndef CONFIG_DISABLE_POLL
 	, 0						/* poll */
 #endif
@@ -71,13 +72,13 @@ static const struct file_operations testcase_drv_fops = {
  ****************************************************************************/
 
 /************************************************************************************
- * Name: testcase_drv_ioctl
+ * Name: kernel_test_drv_ioctl
  *
  * Description:  The standard ioctl method.
  *
  ************************************************************************************/
 
-static int testcase_drv_ioctl(FAR struct file *filep, int cmd, unsigned long arg)
+static int kernel_test_drv_ioctl(FAR struct file *filep, int cmd, unsigned long arg)
 {
 	int ret = -EINVAL;
 	/* Handle built-in ioctl commands */
@@ -184,12 +185,12 @@ static int testcase_drv_ioctl(FAR struct file *filep, int cmd, unsigned long arg
 }
 
 
-static ssize_t testcase_drv_read(FAR struct file *filep, FAR char *buffer, size_t len)
+static ssize_t kernel_test_drv_read(FAR struct file *filep, FAR char *buffer, size_t len)
 {
 	return 0;                                       /* Return EOF */
 }
 
-static ssize_t testcase_drv_write(FAR struct file *filep, FAR const char *buffer, size_t len)
+static ssize_t kernel_test_drv_write(FAR struct file *filep, FAR const char *buffer, size_t len)
 {
 	return len;                                     /* Say that everything was written */
 }
@@ -199,14 +200,14 @@ static ssize_t testcase_drv_write(FAR struct file *filep, FAR const char *buffer
  ****************************************************************************/
 
 /****************************************************************************
- * Name: testcase_drv_register
+ * Name: kernel_test_drv_register
  *
  * Description:
- *   Register /dev/testcase
+ *   Register /dev/kernel_test
  *
  ****************************************************************************/
 
-void testcase_drv_register(void)
+void kernel_test_drv_register(void)
 {
-	(void)register_driver(TESTCASE_DRVPATH, &testcase_drv_fops, 0666, NULL);
+	(void)register_driver(KERNEL_TEST_DRVPATH, &kernel_test_drv_fops, 0666, NULL);
 }
